rename new network to network
