+++ app/src/main/kotlin/io/treehouses/remote/Fragments/DialogFragments/BottomSheetDialogs/EthernetBottomSheet.kt
@@ -6,7 +6,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import io.treehouses.remote.Fragments.NewNetworkFragment
+import io.treehouses.remote.Fragments.NetworkFragment
 import io.treehouses.remote.R
 import io.treehouses.remote.bases.BaseBottomSheetDialog
 import io.treehouses.remote.databinding.DialogEthernetBinding
@@ -22,7 +22,7 @@ class EthernetBottomSheet : BaseBottomSheetDialog() {
             val mask = bind.mask.text.toString()
             listener.sendMessage(getString(R.string.TREEHOUSES_ETHERNET, ip, mask, gateway, dns))
             val intent = Intent()
-            intent.putExtra(NewNetworkFragment.CLICKED_START_CONFIG, true)
+            intent.putExtra(NetworkFragment.CLICKED_START_CONFIG, true)
             targetFragment!!.onActivityResult(targetRequestCode, Activity.RESULT_OK, intent)
             dismiss()
         }
