+++ app/src/main/res/layout/activity_network_fragment.xml
